pyproject.toml
[tool.poetry]
name = "src"
version = "1.0.0"
description = "Real-time sensor data synchronization with CDC and MQTT"
authors = ["shahnoor77"]
[tool.poetry.dependencies]
python = "^3.11"
JPype1 = "^1.4.1"
paho-mqtt = "^1.6.1"
psycopg2-binary = "^2.9.9"
pymongo = "^4.6.1"
mysql-connector-python = "^8.2.0"
pyyaml = "^6.0.1"
cryptography = "^41.0.7"
python-dotenv = "^1.0.0"
prometheus-client = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.12.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"